consolidating course filters + moving to new course types constant - missed in other commit
