--- src/app/(frontend)/embeds/courses/courses-type-filter.tsx
@@ -1,8 +1,0 @@
-'use client'
-
-import { courseTypesData } from '@/collections/Courses/constants'
-import { CheckboxFilter } from '@/components/filters/CheckboxFilter'
-
-export const CoursesTypeFilter = () => {
-  return <CheckboxFilter title="Course Type" urlParam="types" options={courseTypesData} />
-}
