+++ src/collections/Providers/index.ts
@@ -1,4 +1,5 @@
 import { accessByProviderRelationship } from '@/access/byProviderRelationship'
+import { courseTypesData } from '@/constants/courseTypes'
 import { contentHashField } from '@/fields/contentHashField'
 import { stateOptions } from '@/fields/location/states'
 import { slugField } from '@/fields/slug'
@@ -6,7 +7,6 @@ import { validatePhone } from '@/utilities/validatePhone'
 import { validateExternalUrl } from '@/utilities/validateUrl'
 import { validateZipCode } from '@/utilities/validateZipCode'
 import { CollectionConfig } from 'payload'
-import { courseTypesData } from '../Courses/constants'
 import { courseTypesFieldAccess } from './access/courseTypesFieldAccess'
 import { revalidateProvider, revalidateProviderDelete } from './hooks/revalidateProvider'
 import { sendProviderEmails } from './hooks/sendProviderEmails'
